Defined IsFourInDirection()
Implemented functionality for IsFourInDirection() function.

The function checks the next 3 spaces in a given direction. The function is intended to be called after each move and is passed the player's most recent choice and their token. By assuming that the player's most recent choice matches their selected token, the function can omit 1 iteration which is why it iterates 3 instead of 4 times.
Initially the function would have kept track of the numbers of matching tokens found in a count variable, but I omitted that in favor of checking for fail conditions instead. The logic is that there are more conditions where a player does not achieve 4 in a row than there are otherwise. Therefore, the execution can be passed back to the main function quicker by not needing to check the final value of a count variable after all iterations are complete.
Implemented this function for checking to the right.