4 in a row bug fix
Fixed a bug where the the game would not recognize 4 tokens in a row if the last token dropped was in the middle of the 4 tokens

Previous implementation checked for 4 in a row from the last token dropped to the left OR right, up OR down.
Current implementation accurately checks for 4 in a row left AND right, up AND down.

IsFourInDirection -> NumAdjacentTokens

New function returns an int (number of tokens found) instead of a boolean.