Bug Fix
Fixed a bug that allowed tokens on the opposite side of the board to be considered as consecutive.

Found the issue to be related to list comprehension. Instead of going out of bounds when provided board[0][-1], for example, the program would go to the last column in the row, rather than raise an Index Error.