ChoosePlayerToken() Functionality
The player can no choose either 'X' or 'O' as their token.

Post the ChoosePlayerName() Functionality commit, the code is bugged. The main game logic is carried over from the main branch where it assumes the player is player 1.  The player is therefore locked in to play as player 1 even after choosing their name, player # and token.