Bug Fix
Fixed an issue where the game would crash when the AI ran out of moves to make. The issue came from the AI never being allowed to make moves that would result in a loss, even if there are no favorable moves left to make. The AI now chooses a random unfavorable move when there are no favorable moves to make.