Bug Fix
Previous bug that was said to be fixed was not squashed entirely. The problem with the game crashing has been fixed. The issue occurred after setting the default column choice to None when it should be set to a random choice of the current valid moves.

Also added conditionals to improve the best moves the AI makes